Implemented grotesque hack to change 'lane-miles' captions and units to 'track-miles' for project #1570 (Green Line Extension.)
The correct way to do this would be for a field to be added to the tip_projects table indicating if a project is a rail project, and for projects for which this is true, change the captions and units from 'lane-miles' to 'track-miles. This would require a database schema change, which is out of the question at the moment. This is a TBD for the next cycle of work on the database design and app. Sigh.